bug:fixes

+++ components/Payment.tsx
@@ -96,7 +96,7 @@ const Payment = ({
                   destination_latitude: destinationLatitude,
                   destination_longitude: destinationLongitude,
                   ride_time: rideTime.toFixed(0),
-                  fare_price: parseInt(amount) * 100,
+                  fare_price: parseFloat(amount) * 100,
                   payment_status: "paid",
                   driver_id: driverId,
                   user_id: userId,
